==> analysis/02_new_model_input/02.3_google_building_data.ipynb <==
# Google Building Footprint Data

The data is gotten from `https://sites.research.google/open-buildings/#download`

%load_ext jupyter_black
import pandas as pd
import geopandas as gpd
import os
from pathlib import Path
import requests

base_url = (
    Path(os.getenv("STORM_DATA_DIR"))
    / "analysis/02_new_model_input/02_housing_damage/"
)
input_dir = base_url / "input/Google Footprint Data/"
shp_input_dir = base_url / "input/"
output_dir = base_url / "output/"

adm3_shp = gpd.read_file(
    shp_input_dir / "phl_adminboundaries_candidate_adm3.zip"
)
grid = gpd.read_file(output_dir / "phl_0.1_degree_grid_land_overlap.gpkg")
adm3_grid_intersection = gpd.read_file(
    shp_input_dir / "phl_adm3_grid_intersection.gpkg"
)

The files that cover the Philippines are below and are selected by viewing on the map.

### Downloading data

file_pattern = ["33d", "33b", "339", "325", "32f", "323", "331", "347"]
polygons_url_link = "https://storage.googleapis.com/open-buildings-data/v2/polygons_s2_level_4_gzip/"
points_url_link = "https://storage.googleapis.com/open-buildings-data/v2/points_s2_level_4_gzip/"
file_list = [patt + "_buildings.csv.gz" for patt in file_pattern]

Using the points data as we are using centroids.

for file in file_list:
    r = requests.get(points_url_link + file, allow_redirects=True)
    open(input_dir / file, "wb").write(r.content)

### Merging all files into one

google_df = pd.DataFrame()
for file in file_list:
    zone_file = pd.read_csv(input_dir / file, compression="gzip")
    google_df = pd.concat([google_df, zone_file])
google_df

google_df.to_csv(input_dir / "google_footprint_data.csv", index=False)
# google_df = pd.read_csv(input_dir / "google_footprint_data.csv")

google_df[
    (google_df["latitude"].between(17.35, 17.45))
    & (google_df["longitude"].between(122.05, 122.15))
]

# converting point df to geodataframe
ggl_gdf = gpd.GeoDataFrame(
    google_df,
    geometry=gpd.points_from_xy(google_df.longitude, google_df.latitude),
)

ggl_gdf.set_crs(adm3_shp.crs, inplace=True)

# ggl_gdf.to_file(input_dir / "google_footprint_data.gpkg", driver="GPKG")
# ggl_gdf = gpd.read_file(input_dir / "google_footprint_data.gpkg")

## Counting number of buildings by region

del google_df

# ensuring all buildings are within municipality bounds
ggl_gdf_within = gpd.sjoin(ggl_gdf, adm3_shp, how="inner", predicate="within")

del ggl_gdf

ggl_gdf_within.drop(
    ggl_gdf_within.columns.difference(["latitude", "longitude", "geometry"]),
    axis=1,
    inplace=True,
)
ggl_gdf_within["random_id"] = ggl_gdf_within.reset_index().index + 1

ggl_gdf_within[
    (ggl_gdf_within["latitude"].between(17.35, 17.45))
    & (ggl_gdf_within["longitude"].between(122.05, 122.15))
]

ggl_gdf_within.shape

### By Municipality

phl_ggl_bld_municip_count = (
    adm3_shp.sjoin(ggl_gdf_within, how="left")
    .groupby("ADM3_PCODE")
    .count()
    .rename(columns={"random_id": "numbuildings"})
    .reset_index()[["ADM3_PCODE", "numbuildings"]]
)

phl_ggl_bld_municip_count["numbuildings"].sum()

phl_ggl_bld_municip_count.to_csv(
    input_dir / "phl_google_bld_municip_count.csv", index=False
)

del phl_ggl_bld_municip_count

### By Grid

phl_ggl_bld_grid_count = (
    grid.sjoin(ggl_gdf_within, how="left")
    .groupby(["id", "Centroid"])
    .count()
    .rename(columns={"random_id": "numbuildings"})
    .reset_index()[["id", "Centroid", "numbuildings"]]
)

phl_ggl_bld_grid_count["numbuildings"].sum()

phl_ggl_bld_grid_count[phl_ggl_bld_grid_count["Centroid"] == "122.1E_17.4N"]

phl_ggl_bld_grid_count.to_csv(
    input_dir / "phl_google_bld_grid_count.csv", index=False
)

del phl_ggl_bld_grid_count

### By Municipality and Grid Intersection

del adm3_shp, grid

phl_ggl_bld_intersection_count = (
    adm3_grid_intersection.sjoin(ggl_gdf_within, how="left")
    .groupby(["ADM3_PCODE", "id", "Centroid"])
    .count()
    .rename(columns={"random_id": "numbuildings"})
    .reset_index()[["ADM3_PCODE", "id", "Centroid", "numbuildings"]]
)

phl_ggl_bld_intersection_count["numbuildings"].sum()

phl_ggl_bld_intersection_count[
    phl_ggl_bld_intersection_count["Centroid"] == "122.1E_17.4N"
]

phl_ggl_bld_intersection_count[
    ["ADM3_PCODE", "id", "Centroid", "numbuildings"]
].to_csv(input_dir / "phl_google_bld_intersection_count.csv", index=False)
