==> analysis/02_new_model_input/05.0_relative_wealth_index.ipynb <==
# Relative Wealth Index

The data is from `https://data.humdata.org/dataset/relative-wealth-index`

- For the Philippines `https://data.humdata.org/dataset/76f2a2ea-ba50-40f5-b79c-db95d668b843/resource/62233330-20a8-4623-ad87-429b22409ed5/download/phl_relative_wealth_index.csv`

%load_ext jupyter_black
import pandas as pd
import geopandas as gpd
from pathlib import Path
import os

base_dir = Path(os.getenv("STORM_DATA_DIR")) / "analysis/02_new_model_input/"
input_dir = base_dir / "05_vulnerablility/input/"
shp_input_dir = base_dir / "02_housing_damage/output/"
output_dir = base_dir / "05_vulnerablility/output/"

phl_rwi = pd.read_csv(input_dir / "phl_relative_wealth_index.csv")
grid = gpd.read_file(shp_input_dir / "phl_0.1_degree_grid_land_overlap.gpkg")

# creating a geometry column for the point data
phl_rwi_gdf = gpd.GeoDataFrame(
    phl_rwi, geometry=gpd.points_from_xy(phl_rwi.longitude, phl_rwi.latitude)
)

grid_rwi_gdf = gpd.tools.sjoin(phl_rwi_gdf, grid, how="left")
grid_rwi_gdf

# this is used as a check
grid_rwi_gdf[grid_rwi_gdf["Centroid"] == "121.0E_9.6N"]

grid_rwi = pd.DataFrame(
    grid_rwi_gdf.groupby(["id", "Centroid"])["rwi"].mean()
).reset_index()
grid_rwi

grid_rwi.to_csv(output_dir / "phl_rwi_bygrid.csv", index=False)
