==> analysis/02_new_model_input/06.0_ghs_rural_urban_pop.ipynb <==
# GHS Data Download and Aggregating

This notebook describes the process of downloading and aggregating GHS data from [here](https://ghsl.jrc.ec.europa.eu/download.php?). The notebook contains aggregation for the grid:

- Classification of pixels as rural or urban for the raster.
    - Using 21 and above for urban and 13 and below for rural.
- Fraction of each grid pixel that is rural/ urban
- Find population by grid pixel.

The epoch used in closest to the date `27th Feb 2023` and would be 2025.

%load_ext jupyter_black
import pandas as pd
import geopandas as gpd
from pathlib import Path
import os
import requests, zipfile, io
import rasterio
import rioxarray as rxr

from rasterstats import zonal_stats

base_dir = Path(os.getenv("STORM_DATA_DIR")) / "analysis/02_new_model_input/"
input_dir = base_dir / "06_settlement/input/"
shp_input_dir = base_dir / "02_housing_damage/input/"
grid_input_dir = base_dir / "02_housing_damage/output/"
output_dir = base_dir / "06_settlement/output/"

adm3_shp = gpd.read_file(
    shp_input_dir / "phl_adminboundaries_candidate_adm3.zip"
)

# grid
grid = gpd.read_file(grid_input_dir / "phl_0.1_degree_grid_land_overlap.gpkg")

smod_link = "https://jeodpp.jrc.ec.europa.eu/ftp/jrc-opendata/GHSL/GHS_SMOD_GLOBE_R2022A/GHS_SMOD_P2025_GLOBE_R2022A_54009_1000/V1-0/GHS_SMOD_P2025_GLOBE_R2022A_54009_1000_V1_0.zip"
pop_link = "https://jeodpp.jrc.ec.europa.eu/ftp/jrc-opendata/GHSL/GHS_POP_GLOBE_R2022A/GHS_POP_P2025_GLOBE_R2022A_54009_100/V1-0/tiles/"

## Degree of Urbanisation

### Urban-Rural Classification

# Downloading the whole global data set as it is small
req = requests.get(smod_link, verify=False, stream=True)
with zipfile.ZipFile(io.BytesIO(req.content)) as zObj:
    fileNames = zObj.namelist()
    for fileName in fileNames:
        if fileName.endswith("tif"):
            content = zObj.open(fileName).read()
            open(input_dir / "SMOD" / fileName, "wb").write(content)

# Reading in raster
file_name = os.listdir(input_dir / "SMOD")
smod_raster = rasterio.open(input_dir / "SMOD" / file_name[0])
smod_array = smod_raster.read(1)
smod_array

# no data are set at -200
# water seems to be set to 10
# converting to similar CRS
smod_raster.crs

grid.total_bounds

# converting crs and clipping
# checking if crs are the same
smod_raster.crs == grid.crs

smod_raster = rxr.open_rasterio(input_dir / "SMOD" / file_name[0])
smod_raster.rio.crs

smod_raster_wgs84 = smod_raster.rio.reproject(grid.crs)
smod_raster_wgs84_clip = smod_raster_wgs84.rio.clip_box(*grid.total_bounds)

smod_grid_vals = pd.DataFrame(
    {
        "id": grid["id"],
        "Centroid": grid["Centroid"],
        "urban": None,
        "rural": None,
        "water": None,
    }
)
for grd in grid.Centroid:
    grd_sel = grid[grid.Centroid == grd]
    grid_rast = smod_raster_wgs84_clip.rio.clip(
        grd_sel["geometry"], all_touched=False
    )
    smod_grid_vals.loc[grd_sel.index.values, ["urban"]] = (
        (grid_rast >= 21) & (grid_rast <= 30)
    ).sum().values / grid_rast.count().values
    smod_grid_vals.loc[grd_sel.index.values, ["rural"]] = (
        (grid_rast >= 11) & (grid_rast <= 13)
    ).sum().values / grid_rast.count().values
    smod_grid_vals.loc[grd_sel.index.values, ["water"]] = (
        grid_rast == 10
    ).sum().values / grid_rast.count().values

smod_grid_vals.tail(10)

smod_grid_vals[["urban", "rural", "water"]].sum(axis=1).describe()

del (
    smod_raster,
    smod_array,
    smod_raster_wgs84,
    smod_raster_wgs84_clip,
)

## Population

### Total Population by grid

# downloading the popoulation data
# selected from here: https://ghsl.jrc.ec.europa.eu/download.php?ds=pop
phl_boxes = ["R7_C30", "R7_C31", "R8_C30", "R8_C31", "R9_C30", "R9_C31"]
file_list = [
    "GHS_POP_P2025_GLOBE_R2022A_54009_100_V1_0_" + patt + ".zip"
    for patt in phl_boxes
]

for file in file_list:
    req = requests.get(pop_link + file, allow_redirects=True)
    with zipfile.ZipFile(io.BytesIO(req.content)) as zObj:
        fileNames = zObj.namelist()
        for fileName in fileNames:
            if fileName.endswith("tif"):
                content = zObj.open(fileName).read()
                open(input_dir / "POP" / fileName, "wb").write(content)

This section took long to process.

NOTE: 
- The raster layers and the grid do not have the same CRS.
    - Options are to project grid to raster CRS or raster to grid CRS.
    - Projecting to grid CRS seems to produce realistic numbers.
    - Numbers still seem slightly distorted based on visual inspection on QGIS.
    - Example: grid `117.0E_8.2N` 
        - shows a sum of 197.34 visually in QGIS
        - shows a sum of 197.34 after re-projecting the grid to raster CRS.
        - shows a sum of 201.343014 when re-projecting the rasters individually.
        - shows a sum of 355.782260 when merging and re-projecting
        - re-projecting the grid seems the way to go.
- Merging raster layers into one takes a lot of storage. 
    - It also results in higher values.
- A re-projected raster layer takes even more space and is better to just re-project on the fly.

SOLUTION: Re-project grid to raster CRS
The sum of values as total population in grid.
Each raster is computed separately.

# merging rasters using gdal
#! gdalbuildvrt "%STORM_DATA_DIR%/analysis/02_new_model_input/06_settlement/input/POP/PHL_GHS_POP_P2025_R2022A_54009_100_V1_0.vrt" "%STORM_DATA_DIR%/analysis/02_new_model_input/06_settlement/input/POP/GHS*.tif"
# translating virtual mosaic to geotiff
#! gdal_translate -of GTiff -co "TILED=YES" "%STORM_DATA_DIR%/analysis/02_new_model_input/06_settlement/input/POP/PHL_GHS_POP_P2025_R2022A_54009_100_V1_0.vrt" "%STORM_DATA_DIR%/analysis/02_new_model_input/06_settlement/input/POP/PHL_GHS_POP_P2025_R2022A_54009_100_V1_0.tif"

# opening files and merging them
pop_grid_vals = pd.DataFrame(
    {
        "id": grid["id"],
        "Centroid": grid["Centroid"],
        "total_pop": None,
    }
)

## SOLUTION: convert grid to raster CRS

file_list = os.listdir(input_dir / "POP")
tif_list = [tif for tif in file_list if tif.endswith(".tif")]
pop_raster = rasterio.open(input_dir / "POP" / tif_list[0])
grid_crs = grid.to_crs(pop_raster.crs.to_dict())

# looping over tif files
for file in tif_list:
    pop_raster = rasterio.open(input_dir / "POP" / file)
    pop_array = pop_raster.read(1)
    pop_stats = zonal_stats(
        grid_crs,
        pop_array,
        stats=["sum"],
        nodata=-200,
        all_touched=False,
        affine=pop_raster.transform,
    )
    grid_stats = pd.DataFrame(pop_stats)
    pop_grid_vals[phl_boxes[tif_list.index(file)]] = grid_stats["sum"]

# takes too long
# tif_list = os.listdir(input_dir / "POP")
# for file in tif_list[2:3]:
#    pop_rast = (
#        rxr.open_rasterio(input_dir / "POP" / file)
#        # .rio.reproject(grid.crs)
#        # .rio.clip_box(*grid.total_bounds)
#    )
#    for grd in grid_crs.Centroid[0:10]:
#        grd_sel = grid_crs[grid_crs.Centroid == grd]
#        try:
#            grid_rast = pop_rast.rio.clip(
#                grd_sel["geometry"], all_touched=True
#            )
#            pop_grid_vals.loc[
#                grd_sel.index.values, [phl_boxes[tif_list.index(file)]]
#            ] = ((grid_rast.where(grid_rast >= 0)).sum().values)
#        except:
#            pop_grid_vals.loc[
#                grd_sel.index.values, [phl_boxes[tif_list.index(file)]]
#            ] = 0

# sum all columns
pop_grid_vals["total_pop"] = pop_grid_vals.loc[:, phl_boxes].sum(axis=1)

pop_grid_vals

pop_grid_vals["total_pop"].sum()

# merging the two dataframes
merged_ghs_df = smod_grid_vals.merge(
    pop_grid_vals[["id", "Centroid", "total_pop"]], on=["id", "Centroid"]
)

merged_ghs_df

# writing output
merged_ghs_df.to_csv(output_dir / "ghs_rural_urban_pop.csv", index=False)
